+++ src/test/java/com/google/javascript/cl2dts/closure.depgraph
@@ -1,148 +1,7 @@
-[
-  [
-    "nonroots",
-    [
-      [
-        "javascript/closure/base.js",
-        [
-          [
-            "custom_annotations",
-            []
-          ],
-          [
-            "is_externs",
-            false
-          ],
-          [
-            "is_module",
-            false
-          ],
-          [
-            "mod_names",
-            [
-              "$$DEFAULT"
-            ]
-          ],
-          [
-            "modifies",
-            []
-          ],
-          [
-            "provides",
-            [
-              "$$$javascript/closure/base.js",
-              "goog"
-            ]
-          ],
-          [
-            "requires",
-            []
-          ]
-        ]
-      ]
-    ]
-  ],
-  [
-    "resource_roots",
-    []
-  ],
-  [
-    "roots",
-    [
-      [
-        "my/thing/static/js/0-bootstrap.js",
-        [
-          [
-            "custom_annotations",
-            []
-          ],
-          [
-            "is_externs",
-            false
-          ],
-          [
-            "is_module",
-            false
-          ],
-          [
-            "mod_names",
-            [
-              "$$DEFAULT"
-            ]
-          ],
-          [
-            "modifies",
-            []
-          ],
-          [
-            "provides",
-            [
-              "$$$my/thing/static/js/0-bootstrap.js",
-              "foo.application.module"
-            ]
-          ],
-          [
-            "requires",
-            [
-              "foo.annotations.module",
-              "foo.comments.module",
-              "foo.constants.module",
-              "foo.exceptions.module",
-              "foo.experimental.module",
-              "foo.filters.module",
-              "foo.general.module",
-              "foo.motionVideo.module",
-              "foo.stats.module",
-              "foo.toast.module",
-              "foo.users.module",
-              "foo.utils.module",
-              "goog"
-            ]
-          ]
-        ]
-      ],
-      [
-        "my/thing/static/js/annotations/annotations-canvas-controller.js",
-        [
-          [
-            "custom_annotations",
-            []
-          ],
-          [
-            "is_externs",
-            false
-          ],
-          [
-            "is_module",
-            false
-          ],
-          [
-            "mod_names",
-            [
-              "$$DEFAULT"
-            ]
-          ],
-          [
-            "modifies",
-            []
-          ],
-          [
-            "provides",
-            [
-              "$$$my/thing/static/js/annotations/annotations-canvas-controller.js",
-              "foo.annotations.AnnotationEvent",
-              "foo.annotations.AnnotationsCanvasController",
-              "foo.annotations.AnnotationsCanvasEvent"
-            ]
-          ],
-          [
-            "requires",
-            [
-              "goog"
-            ]
-          ]
-        ]
-      ]
-    ]
-  ]
+[ [ "nonroots", [ [ "javascript/closure/base.js"] ]],
+  [ "resource_roots", [] ],
+  [ "roots", [
+    [ "my/thing/static/js/0-bootstrap.js"],
+    [ "my/thing/static/js/annotations/annotations-canvas-controller.js"]
+  ] ]
 ]
